Move catch.hpp to separate folder
This is to avoid direct inclusion of catch.hpp when use_catch.h should be used
instead. The first guard here is module_dependencies.txt, where only
testing-utils must have "catch" listed. The second line of defense is a macro
that must be set.
