+++ unit/catch/CMakeLists.txt
@@ -1,0 +1,4 @@
+target_include_directories(catch
+    PUBLIC
+    ${CMAKE_CURRENT_SOURCE_DIR}/..
+)
