+++ unit/catch/catch.hpp
@@ -8,6 +8,11 @@
  *  Distributed under the Boost Software License, Version 1.0. (See accompanying
  *  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
  */
+
+#ifndef INCLUDED_VIA_USE_CATCH_H
+#error Do not include this file directly, use use_catch.h instead
+#endif
+
 #ifndef TWOBLUECUBES_SINGLE_INCLUDE_CATCH_HPP_INCLUDED
 #define TWOBLUECUBES_SINGLE_INCLUDE_CATCH_HPP_INCLUDED
