+++ unit/testing-utils/module_dependencies.txt
@@ -1,4 +1,5 @@
 ansi-c
+catch
 testing-utils
 util
 analyses
